third_party/blink/web_tests/external/wpt/css/css-text-decor/parsing/text-decoration-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['text-decoration'] = "solid" should set the property value
  assert_equals: serialization should be canonical expected "none" but got "solid"
[FAIL] e.style['text-decoration'] = "currentcolor" should set the property value
  assert_equals: serialization should be canonical expected "none" but got "currentcolor"
Harness: the test ran to completion.
